# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "Ack": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "Ack",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Every error body. `code` is stable; clients translate it.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicUserOut": {
        "description": "What other users may see. Never includes the phone number.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identity_verified": {
            "title": "Identity Verified",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "name",
          "identity_verified",
          "created_at"
        ],
        "title": "PublicUserOut",
        "type": "object"
      },
...
      "ShowroomCreate": {
        "properties": {
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "commercial_registration_number": {
            "maxLength": 20,
            "minLength": 6,
            "title": "Commercial Registration Number",
            "type": "string"
          },
          "name_ar": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name Ar",
            "type": "string"
          },
          "name_en": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          }
        },
        "required": [
          "name_ar",
          "commercial_registration_number",
          "city_id"
        ],
        "title": "ShowroomCreate",
        "type": "object"
      },
      "ShowroomOut": {
        "properties": {
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name_ar": {
            "title": "Name Ar",
            "type": "string"
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          },
          "subscription_tier": {
            "$ref": "#/components/schemas/SubscriptionTier"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name_ar",
          "name_en",
          "city_id",
          "verified",
          "subscription_tier"
        ],
        "title": "ShowroomOut",
        "type": "object"
      },
      "ShowroomStaffAdd": {
        "properties": {
          "phone": {
            "maxLength": 20,
            "minLength": 6,
            "title": "Phone",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ShowroomStaffRole",
            "default": "agent"
          }
        },
        "required": [
          "phone"
        ],
        "title": "ShowroomStaffAdd",
        "type": "object"
      },
      "ShowroomStaffOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ShowroomStaffRole"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserOut"
          }
        },
        "required": [
          "id",
          "role",
          "user"
        ],
        "title": "ShowroomStaffOut",
        "type": "object"
      },
      "ShowroomStaffRole": {
        "enum": [
          "owner",
          "manager",
          "agent"
        ],
        "title": "ShowroomStaffRole",
        "type": "string"
      },
      "ShowroomUpdate": {
        "properties": {
          "city_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "name_ar": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Ar"
          },
          "name_en": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          }
        },
        "title": "ShowroomUpdate",
        "type": "object"
      },
      "SubscriptionTier": {
        "enum": [
          "free",
          "basic",
          "pro"
        ],
        "title": "SubscriptionTier",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /auth/verify",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "[APP_NAME] API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/showrooms": {
      "post": {
        "operationId": "create_showroom",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowroomCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowroomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register a showroom (the creator becomes its owner)",
        "tags": [
          "showrooms"
        ]
      }
    },
    "/showrooms/{showroom_id}": {
      "get": {
        "operationId": "get_showroom",
        "parameters": [
          {
            "in": "path",
            "name": "showroom_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Showroom Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowroomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Showroom",
        "tags": [
          "showrooms"
        ]
      },
      "patch": {
        "operationId": "update_showroom",
        "parameters": [
          {
            "in": "path",
            "name": "showroom_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Showroom Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowroomUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowroomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Showroom",
        "tags": [
          "showrooms"
        ]
      }
    },
    "/showrooms/{showroom_id}/staff": {
      "get": {
        "operationId": "list_showroom_staff",
        "parameters": [
          {
            "in": "path",
            "name": "showroom_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Showroom Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShowroomStaffOut"
                  },
                  "title": "Response List Showroom Staff",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Staff list (visible to the showroom's own staff)",
        "tags": [
          "showrooms"
        ]
      },
      "post": {
        "operationId": "add_showroom_staff",
        "parameters": [
          {
            "in": "path",
            "name": "showroom_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Showroom Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowroomStaffAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowroomStaffOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add a staff member by phone number",
        "tags": [
          "showrooms"
        ]
      }
    },
    "/showrooms/{showroom_id}/staff/{staff_id}": {
      "delete": {
        "operationId": "remove_showroom_staff",
        "parameters": [
          {
            "in": "path",
            "name": "showroom_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Showroom Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "staff_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Staff Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Showroom Staff",
        "tags": [
          "showrooms"
        ]
      }
    },
    "/showrooms/{showroom_id}/verify": {
      "post": {
        "operationId": "verify_showroom",
        "parameters": [
          {
            "in": "path",
            "name": "showroom_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Showroom Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowroomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark a showroom as verified (admin only)",
        "tags": [
          "showrooms"
        ]
      }
    },
...
  }
}